Update components-slots.md
The previously used grammar "from the child" was ambiguous, because it can be interpreted as "of the child" vs "from within the child." Users that are new to slots, templates, etc would spend time deciphering whether we mean data that is in the child vs data that is outside the child, but accessed from within the child. So, I re-wrote parts of this document to make this distinction crystal clear.